requirements.txt
Django==5.1.6
djangorestframework==3.15.2
openai==1.64.0
python-dotenv==1.0.1
